% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_explorer.R
\name{trace_explorer}
\alias{trace_explorer}
\alias{plotly_trace_explorer}
\title{Trace explorer}
\usage{
trace_explorer(eventlog, coverage = NULL, n_traces = NULL,
  min_trace_coverage = NULL, type = c("frequent", "infrequent"),
  .abbreviate = T, show_labels = T,
  scale_fill = scale_fill_discrete(h = c(0, 360) + 15, l = 40),
  raw_data = F, frequency = c("relative", "absolute"))

plotly_trace_explorer(eventlog, coverage = NULL, n_traces = NULL,
  type = c("frequent", "infrequent"), .abbreviate = T,
  show_labels = T, scale_fill = scale_fill_discrete(h = c(0, 360) + 15,
  l = 40), raw_data = F)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{coverage}{The percentage coverage of the trace to explore.}

\item{n_traces}{Instead of setting coverage, you can set an exact number of traces. Should be an integer larger than 0.}

\item{min_trace_coverage}{The minimum coverage a trace has to be considered explore.}

\item{type}{Frequent or infrequenct traces to explore}

\item{.abbreviate}{If TRUE, abbreviate activity labels}

\item{show_labels}{If False, activity labels are not shown.}

\item{scale_fill}{Set color scale}

\item{raw_data}{Return raw data}

\item{frequency}{Display 'relative' or 'absolute' values on plot.}
}
\description{
Explore traces, ordered by relative trace frequency
}
